Changed dlploy_argon_lmps_generator to allow changing density, cutoff and random number seed from launch options.
The above options can now be chnaged from the command line with
the R,C and S flags. Options required to recreate a config are
stored in a SIMULATION_RECORD file and can be used to recreate
the configs with the F flag. This avoids storing potentially
large config files.