Fixed issues arising from particles on boundaries.
BC conditionals are now >= on upper boundaries and < on lower
boundaries. Cell binning (on CPU) has been moved to use the domain
inner boundary with the resulting bin incremented by one in each
direction. The BC code now catches particles sitting on the upper
boundary and moves them to the lower boundary of the adjacent
process.

In previous code the lower boundary was compared with the outer
process boundary. If a particle was sat exactly on the boundary
rounding would place the particle in a halo cell. Starting at zero
with int round towards zero solves the case where a particle has
a position which rounds to be slightly negative.